change the database structure
