+++ MainApp.js
@@ -2,27 +2,27 @@ import React from "react";
 import { NavigationContainer } from "@react-navigation/native";
 import { createStackNavigator } from "@react-navigation/stack";
 import { StyleSheet, View, Text, TextInput, Button, Alert } from "react-native";
-import FCOwnerScreen from "./screens/FCOwnerScreen";
-import FCStallOwnerScreen from "./screens/FCStallOwnerScreen";
+import FCOwnerScreen from "./screens/foodCentreOwner/FCOwnerScreen";
+import FCStallOwnerScreen from "./screens/stallOwner/FCStallOwnerScreen";
 import { AppLoading } from "expo";
 import { Asset } from "expo-asset";
 import { Font } from "expo-font";
-import PatronFoodCentre from "./screens/PatronFoodCentre";
-import FoodCentreHome from "./screens/FoodCentreHome";
-import PatronSeat from "./screens/PatronSeat";
-import PatronStall from "./screens/PatronStall";
-import StallMenu from "./screens/StallMenu";
+import PatronFoodCentre from "./screens/patron/PatronFoodCentre";
+import FoodCentreHome from "./screens/patron/FoodCentreHome";
+import PatronSeat from "./screens/patron/PatronSeat";
+import PatronStall from "./screens/patron/PatronStall";
+import StallMenu from "./screens/patron/StallMenu";
 import LoginScreen from "./screens/auth/LoginScreen";
 import SignUpScreen from "./screens/auth/SignUpScreen";
 import ForgotPasswordScreen from "./screens/auth/ForgotPasswordScreen";
-import AddFCScreen from "./screens/AddFCScreen";
-import FCOwnerPersonalList from "./screens/FCOwnerPersonalList";
-import EditFCScreen from "./screens/EditFCScreen";
-import AddStallNMenuScreen from "./screens/AddStallNMenuScreen";
-import FCStallPersonalList from "./screens/FCStallPersonalList";
-import FCStallMenuPersonalList from "./screens/FCStallMenuPersonalList";
-import EditStallScreen from "./screens/EditStallScreen";
-import EditMenuScreen from "./screens/EditMenuScreen";
+import AddFCScreen from "./screens/foodCentreOwner/AddFCScreen";
+import FCOwnerPersonalList from "./screens/foodCentreOwner/FCOwnerPersonalList";
+import EditFCScreen from "./screens/foodCentreOwner/EditFCScreen";
+import AddStallNMenuScreen from "./screens/stallOwner/AddStallNMenuScreen";
+import FCStallPersonalList from "./screens/stallOwner/FCStallPersonalList";
+import FCStallMenuPersonalList from "./screens/stallOwner/FCStallMenuPersonalList";
+import EditStallScreen from "./screens/stallOwner/EditStallScreen";
+import EditMenuScreen from "./screens/stallOwner/EditMenuScreen";
 import {
   FOODCENTRE_USER,
   STALL_USER,
