+++ app-redux/historyActions.js
@@ -1,4 +1,4 @@
-import db from "../firebase/firebaseConfig";
+import { db } from "../firebase/firebaseConfig";
 import * as firebase from "firebase";
 import {
   FOODCENTRE_USER,
