+++ package.json
@@ -33,7 +33,9 @@
     "react-native-vector-icons": "^6.6.0",
     "react-native-web": "~0.11.7",
     "react-redux": "^7.2.0",
+    "react-redux-firebase": "^3.6.0",
     "redux": "^4.0.5",
+    "redux-firestore": "^0.13.0",
     "redux-thunk": "^2.3.0"
   },
   "devDependencies": {
