+++ screens/stallOwner/FCStallMenuPersonalList.js
@@ -7,10 +7,12 @@ import {
   TouchableOpacity,
   TextInput,
 } from "react-native";
-import { globalstyles } from "../styles/globalstyles";
+import { globalstyles } from "../../styles/globalstyles";
 import { MaterialIcons, AntDesign } from "@expo/vector-icons";
 import { connect } from "react-redux";
-import { deleteMenusData } from "../app-redux/actions";
+import { deleteMenusData } from "../../app-redux/actions";
+import { firestoreConnect } from "react-redux-firebase";
+import { compose } from "redux";
 
 function FCStallMenuPersonalList(props) {
   const uid = props.user.userId;
@@ -53,7 +55,7 @@ function FCStallMenuPersonalList(props) {
                   name="delete"
                   size={30}
                   style={styles.icon}
-                  onPress={() => props.deleteMenusData(item)}
+                  onPress={() => props.deleteMenusData(item.id)}
                 />
                 <AntDesign
                   name="edit"
@@ -106,7 +108,7 @@ const styles = StyleSheet.create({
 
 const mapStateToProps = (state) => ({
   user: state.user,
-  menus: state.menus,
+  menus: state.firestore.ordered.menus,
 });
 
 const mapDispatchToProps = (dispatch) => {
@@ -115,7 +117,7 @@ const mapDispatchToProps = (dispatch) => {
   };
 };
 
-export default connect(
-  mapStateToProps,
-  mapDispatchToProps
+export default compose(
+  connect(mapStateToProps, mapDispatchToProps),
+  firestoreConnect([{ collection: "menus", orderBy: ["name", "asc"] }])
 )(FCStallMenuPersonalList);
